Move functional graph creation to testing utils
ghstack-source-id: 56c64dcd841e45999ba032468f619a2f7a9010e9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34694
